+++ samples/detectNunit.js
@@ -3,23 +3,6 @@
 const path = require('path');
 const vsUtils = require('../src');
 
-const determineNunitRunner = (semver) => {
-  if(parseInt(semver.major, 10) >= 3) {
-    return `NUnit.Runners.${semver.version}`;
-  } else {
-    return `NUnit.ConsoleRunner.${semver.version}`;
-  }
-};
-
-const determineNunitExecutable = (semver, arch) => {
-  if(parseInt(semver.major, 10) >= 3) {
-    const archFlag = (arch && arch.toUpperCase() === 'X86') ? '-x86' : '';
-    return `nunit-console${archFlag}.exe`;
-  } else {
-    return 'nunit3-console.exe';
-  }
-};
-
 const slnFile = path.join(__dirname, './TestConsoleApplication/TestConsoleApplication.sln');
 
 vsUtils.parseSolution(slnFile).then(solution => {